use try...catch for sqlite commands

+++ engine/note/delete.php
@@ -4,14 +4,21 @@ require_once 'engine.php';
 
 load_db();
 
-$result = $db->exec("DELETE FROM notes WHERE title = '$title';");
-if ($result) {
-  $messageType = "alert-success";
-  $message = "Note <b>$title</b> supprimée";
-  $title = "";
+$note_exists = $db->querySingle(
+  "SELECT EXISTS(SELECT title FROM notes WHERE title = '$title');"
+);
+if($note_exists) {
+  try {
+    $result = $db->exec("DELETE FROM notes WHERE title = '$title';");
+    $message = "Note <b>$title</b> supprimée";
+    $title = "";
+  } catch(Throwable $err) {
+    $message = $err.message;
+    $messageType = "alert-danger";
+  }
 } else {
-  $messageType = "alert-danger";
-  $message = "La note <b>$title</b> n'a pas pu être supprimée";
+  $message = "La note $title n'existe pas";
+  $messageType = "alert-warning";
 }
 
 require_once 'read.php';
